JS APIv3 and error fixes
Updated to FME Server JS API v3
Commented out the abililty to select WGS84 geocentric as a destination coordinate system as this resulted in job failure.
Added additional error handling and error messaging when FME Server responds with a 422 error. This commonly appears to be when the generic reader can't correctly identify source file and open a correct reader.
Added additional error handling and error messging when the demo returns a download link but no features were written, resulting in a 404.